+++ src/core/index.ts
@@ -1,2 +1,2 @@
-export * from "./dwn.js";
-export * from "./server.js";
+export * from "./dcx-dwn.js";
+export * from "./dcx-server.js";
