+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './config/index.js';
+export * from './config/env-config.js';
 export * from './core/index.js';
 export * from './protocol/index.js';
 export * from './utils/index.js';
