+++ src/protocol/index.ts
@@ -1,4 +1,3 @@
-export * from './handlers/dcx.js';
 
 import ExampleManifest from './manifests/EXAMPLE-MANIFEST.json';
 export { ExampleManifest };
@@ -7,6 +6,7 @@ export { schema as presentationSchema } from './schemas/application.js';
 export { schema as responseSchema } from './schemas/response.js';
 export { schema as invoiceSchema } from './schemas/invoice.js';
 export { schema as manifestSchema } from './schemas/manifest.js';
+export * from './handlers.js';
 
 export { protocol as credentialApplicantProtocol } from './credential-applicant.js';
 export { protocol as credentialIssuerProtocol } from './credential-issuer.js';
