+++ src/utils/constants.ts
@@ -1,6 +1,6 @@
 export const WEB5_CONNECT_PASSWORD_WARNING = `SECURITY WARNING: You have not set a WEB5_CONNECT_PASSWORD, one will be generated for you and saved to web5.password`;
 
-export const WEB5_CONNECT_RECOVERY_PHRASE_WARNING = `SECURITY WARNING: You have not set a WEB5_CONNECT_RECOVERY_PHRASE, one will be generated for you and saved to web5.seed`;
+export const WEB5_CONNECT_RECOVERY_PHRASE_WARNING = `SECURITY WARNING: You have not set a WEB5_CONNECT_RECOVERY_PHRASE, one will be generated for you and saved to file web5.seed`;
 
 export const CIPHER_KEY_WARNING = `SECURITY WARNING: You have not set a CIPHER_KEY, one will be generated for you and saved to cipher.key`;
