+++ src/utils/did-manager.ts
@@ -1,61 +1,40 @@
+import { Ed25519, Jwk } from '@web5/crypto';
 import {
-  PortableDid,
   BearerDid,
   DidDht,
   DidDhtCreateOptions,
-  DidDocument,
-  DidDocumentMetadata,
   DidRegistrationResult,
   DidResolutionResult,
+  PortableDid
 } from '@web5/dids';
 import { readFile } from 'fs/promises';
-import { DcxDidError } from './error.js';
-import { Ed25519, Jwk } from '@web5/crypto';
-import { dcxEnvConfig } from '../config/index.js';
+import { dcxEnvConfig } from '../config/env-config.js';
+import { DidManagerConfig, JWK_PRIVATE_KEY_FORMAT } from '../types/did.js';
+import { handleAsyncErrors } from './error.js';
 
-export class PortableDidBuilder implements PortableDid {
-  uri: string;
-  document: DidDocument;
-  metadata: DidDocumentMetadata;
-  privateKeys: Jwk[];
+export class DidManagerBuilder {
+  public did: string;
+  public bearerDid: BearerDid;
+  public portableDid: PortableDid;
 
-  constructor(
-    uri: string,
-    document: DidDocument,
-    metadata: DidDocumentMetadata,
-    privateKeys: Jwk[],
-  ) {
-    this.uri = uri;
-    this.document = document;
-    this.metadata = metadata;
-    this.privateKeys = privateKeys;
+  constructor(config: DidManagerConfig) {
+    this.bearerDid = config.bearerDid;
+    this.portableDid = config.portableDid;
+    this.did = config.did;
   }
 }
-
-export type DidManagerConfig = {
-  did: string;
-  bearerDid: BearerDid;
-  portableDid: PortableDid;
-};
-
-export const JWK_PRIVATE_KEY_FORMAT = { crv: 'Ed25519', kty: 'OKP', x: '' };
-
-export class DidUtil {
+export class DidUtil extends DidManagerBuilder {
   /**
-   *
-   * Uses Ed25519 to generate a private key; see {@link Ed25519.computePublicKey()}
-   * @param privKey the private key to generate the public key from
-   * @returns Jwk; see {@link Jwk}
-   */
-  public static async computeDidJwkPublicKey(privKey: string): Promise<Jwk> {
-    try {
-      const key = { ...JWK_PRIVATE_KEY_FORMAT, d: privKey } as Jwk;
-      const keyPair = await Ed25519.computePublicKey({ key });
-      return { ...keyPair, x: keyPair.x };
-    } catch (error: any) {
-      console.error('computeDidJwkPublicKey', error);
-      throw new DcxDidError('Failed to compute public key');
-    }
+  *
+  * Uses Ed25519 to generate a private key; see {@link Ed25519.computePublicKey()}
+  * @param privKey the private key to generate the public key from
+  * @returns Jwk; see {@link Jwk}
+  */
+  @handleAsyncErrors
+  public static async computeDidJwkPublicKey(privKey: string): Promise<Jwk | void | TypedPropertyDescriptor<any>> {
+    const key = { ...JWK_PRIVATE_KEY_FORMAT, d: privKey } as Jwk;
+    const keyPair = await Ed25519.computePublicKey({ key });
+    return { ...keyPair, x: keyPair.x } as Jwk;
   }
 
   /**
@@ -64,14 +43,9 @@ export class DidUtil {
    * @param options The did dht create options object; see {@link DidDhtCreateOptions}
    * @returns BearerDid; see {@link BearerDid}
    */
-  public async createBearerDid(options: DidDhtCreateOptions<any>): Promise<BearerDid> {
-    try {
-      return await DidDht.create({ options });
-      // const portableDid = await bearerDid.export();
-    } catch (error: any) {
-      console.error('createBearerDid', error);
-      throw new DcxDidError('Failed to create bearerDid');
-    }
+  @handleAsyncErrors
+  public static async createBearerDid(options: DidDhtCreateOptions<any>): Promise<BearerDid> {
+    return await DidDht.create({ options });
   }
 
   /**
@@ -80,42 +54,21 @@ export class DidUtil {
    * @param didUri the uri to resolve
    * @returns DidResolutionResult; see {@link DidResolutionResult}
    */
-  public async resolveDidDoc(didUri: string): Promise<DidResolutionResult> {
-    try {
-      return await DidDht.resolve(didUri);
-    } catch (error) {
-      console.error('resolveDidDoc', error);
-      throw new DcxDidError('Failed to resolve didDocument using didUri');
-    }
+  @handleAsyncErrors
+  public static async resolveDidDoc(didUri: string): Promise<DidResolutionResult> {
+    return await DidDht.resolve(didUri);
   }
 }
 
 export class DidManager extends DidUtil {
-  public did: string;
-  public bearerDid: BearerDid;
-  public portableDid: PortableDid;
-
-  constructor(config: DidManagerConfig) {
-    super();
-    this.bearerDid = config.bearerDid;
-    this.portableDid = config.portableDid;
-    this.did = config.did;
-  }
-
   /**
-   *
-   * @param gatewayUri the uri of the gateway to publish the did to
-   * @returns DidRegistrationResult; see {@link DidRegistrationResult}
-   */
-  public async publishDidDoc(
-    gatewayUri: string = dcxEnvConfig.DHT_GATEWAY_ENDPOINT,
-  ): Promise<DidRegistrationResult> {
-    try {
-      return await DidDht.publish({ did: this.bearerDid, gatewayUri });
-    } catch (error) {
-      console.error('publishDidDoc', error);
-      throw new DcxDidError('Failed to publish did to gateway');
-    }
+ *
+ * @param gatewayUri the uri of the gateway to publish the did to
+ * @returns DidRegistrationResult; see {@link DidRegistrationResult}
+ */
+  @handleAsyncErrors
+  public async publishDidDoc(gatewayUri: string = dcxEnvConfig.DHT_GATEWAY_ENDPOINT): Promise<DidRegistrationResult> {
+    return await DidDht.publish({ did: this.bearerDid, gatewayUri });
   }
 
   /**
@@ -124,16 +77,12 @@ export class DidManager extends DidUtil {
    * @param didFilepath the path to the file containing the portable did object; see {@link PortableDid}
    * @returns BearerDid; see {@link BearerDid}
    */
+  @handleAsyncErrors
   public async importPortableDidFromFile(didFilepath: string): Promise<BearerDid> {
-    try {
-      const didFileString = (await readFile(didFilepath))?.toString();
-      const portableDid = JSON.parse(didFileString);
-      this.portableDid = portableDid;
-      return await this.importPortableDid(portableDid);
-    } catch (error: any) {
-      console.error('importPortableDidFromFile', error);
-      throw new DcxDidError('Failed to import portableDid from didFilepath');
-    }
+    const didFileString = (await readFile(didFilepath))?.toString();
+    const portableDid = JSON.parse(didFileString);
+    this.portableDid = portableDid;
+    return await this.importPortableDid(portableDid);
   }
 
   /**
@@ -141,14 +90,10 @@ export class DidManager extends DidUtil {
    * @param portableDid a portable did object; see {@link PortableDid}
    * @returns BearerDid; see {@link BearerDid}
    */
+  @handleAsyncErrors
   public async importPortableDid(portableDid: PortableDid): Promise<BearerDid> {
-    try {
-      const bearerDid = await DidDht.import({ portableDid: this.portableDid ?? portableDid });
-      this.bearerDid = bearerDid;
-      return bearerDid;
-    } catch (error: any) {
-      console.error('importPortableDid', error);
-      throw new DcxDidError('Failed to import portableDid');
-    }
+    const bearerDid = await DidDht.import({ portableDid: this.portableDid ?? portableDid });
+    this.bearerDid = bearerDid;
+    return bearerDid;
   }
 }
