+++ src/utils/error.ts
@@ -1,36 +1,41 @@
 export class DcxError extends Error {
-  constructor(public message: string) {
-    super(message);
-    this.name = 'DcxError';
+  constructor(public error: any, name: string) {
+    super(error);
+    this.name = name;
   }
 }
 
-export class DcxDidError extends Error {
-  constructor(
-    public message: string,
-    did?: string,
-  ) {
-    super(`${did} - ${message}`);
-    this.name = 'DcxDidError';
+export class DidManagerError extends DcxError {
+  constructor(public error: any) {
+    super(error, 'DidManagerError');
   }
 }
 
-export class DcxServerError extends Error {
-  constructor(
-    public message: string,
-    error?: any,
-  ) {
-    super(error?.message ?? message);
-    this.name = 'DcxServerError';
+export class DcxServerError extends DcxError {
+  constructor(public error: any) {
+    super(error, 'DcxServerError');
   }
 }
 
-export class DcxDwnError extends Error {
-  constructor(
-    public code: number,
-    message: string,
-  ) {
-    super(`${code} - ${message}`);
-    this.name = 'DcxDwnError';
+export class DcxDwnError extends DcxError {
+  constructor(public code: number, message: string) {
+    super(`${code} - ${message}`, 'DcxDwnError');
   }
 }
+
+export function handleAsyncErrors(target: any, propertyKey: any, descriptor?: any): any {
+  if (!descriptor) {
+    descriptor = Object.getOwnPropertyDescriptor(target, propertyKey)!;
+  }
+  const originalMethod = descriptor.value;
+  descriptor.value = async function (...args: any[]) {
+    try {
+      return await originalMethod.apply(this, args);
+    } catch (error) {
+      console.error(`${propertyKey}`, 'Failed', error);
+      throw new DidManagerError(error);
+    }
+  };
+  return descriptor;
+}
+
