+++ test.ts
@@ -1,0 +1,6 @@
+import { DcxServer } from './src/index';
+import CustomManifest from './CUSTOM-MANIFEST.json';
+
+const server = new DcxServer({ needWeb5Init: true });
+server.useManifest('customManifest', CustomManifest);
+server.start();
